Add 'position' attribute to influence the sequence of partitions.
This makes it possible to put vendor partitions in front of the
pre-defined Brillo partitions. This may be needed by some boot loaders.

Also make it permissible to omit the 'type_guid' attribute by defaulting
to 'brillo_vendor_specific' if omitted.

Slightly rework FakeGuidGenerator to ensure that we get stable GUIDs for
A/B partitions when feeding its output back into itself (idempotence).

TEST=New unit test + unit tests pass.
BUG=28252419

Change-Id: I7a8e18449a5f16105f429abf0fb6114a3b358fea