Introduce rpi_boot partition option
A partition layout may specify a single rpi_boot partition. This
partition will show up in the protective mbr alongside the gpt table. We
will also change the dimensions of the MBR GPT protective partition to
prevent it from overlapping with the boot partition (which will crash
the Pi's firmware).

Change-Id: Ic8ec7ce4f06117cc90dc7c04e53aa837b36fb1bc
Test: Modified unit tests pass
Bug: 28915785