Add 'partitions_offset_begin' attribute.
This attribute can set the target disk partitions offset
begin size. It is good for protecting some read-only data.

With this attribute, it makes security to protect SoC
read-only bootloader. To ensure that system boot security,
the SoC read-only bootloader has special format that MUST
use the SoC tool to flash it.

Change-Id: I5102f11ab37e1f706b2a1774ff888c29226a115f
Signed-off-by: Weiwen Chen <cww@rock-chips.com>